Fix warning in VALID_INDEX macro.

+++ code/game/q_shared.h
@@ -281,7 +281,7 @@ typedef enum {
 
 #define ARRAY_LEN(x) (sizeof (x) / sizeof( *(x) ))
 
-#define VALID_INDEX(x,y) (y >= 0 && y < ARRAY_LEN(x))
+#define VALID_INDEX(x,y) (y >= 0 && (size_t)y < ARRAY_LEN(x))
 
 //
 // these aren't needed by any of the VMs.  put in another header?
